+++ App/Entities/ApiPath.cs
@@ -41,48 +41,4 @@ namespace App.Entities
 
         public override string ToString() => raw;
     }
-
-    public class ApiPathPart
-    {
-        string text;
-
-        public ApiPathPart(string text, ApiPath path, ApiPathPart parent, int index)
-        { 
-            this.text = text;
-
-            Path = path;
-            Parent = parent;
-            Index = index;
-        }
-
-        #region Properties
-        public ApiPath Path { get; set; }
-        public ApiPathPart Parent { get; set; }
-        public int Index { get; set; }
-
-        public bool IsIdentifier => text.StartsWith('{') && text.EndsWith('}');
-
-        public bool IsRespectingCase(CASE casePattern)
-        {
-            text = text.Replace("{", "").Replace("}", "");
-
-            if (text.All(char.IsLower))
-                return true;
-
-            switch (casePattern)
-            {
-                case CASE.CAMEL:
-                    return text.All(c => char.IsLetter(c) || char.IsNumber(c));
-                case CASE.SNAKE:
-                    return text.Contains('_') && !text.Contains('-') && !text.Any(c => char.IsUpper(c));
-                case CASE.SPINAL:
-                    return text.Contains('-') && !text.Contains('_') && !text.Any(c => char.IsUpper(c));
-                default:
-                    return true;
-            }
-        }
-        #endregion
-
-        public override string ToString() => text;
-    }
 }
